requirements.txt
pygame~=2.1.2
pygame_gui
PyYAML~=6.0